+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Most recent releases are shown at the top. Each release shows:
 - Uses [Zephyr-7B](https://huggingface.co/TheBloke/zephyr-7B-beta-GGUF) as default model in `webapp.yml`. (#52)
 
 ### changed
-- N/A
+- Added `stop` paramter to `LLM.prompt` (overrides `stop` paramter supplied to constructor) (#53)
 
 ### fixed:
 - N/A